Auth Check

+++ resources/views/book-requests/index.blade.php
@@ -15,13 +15,16 @@
                                     {{ $bookRequest->title . ', ' . $bookRequest->edition }}<br>
                                     by {{ $bookRequest->author }} <br>
 
-                                    @if ( auth()->user()->id  != $bookRequest->user->id)
-                                        <a href="/book-requests/{{ $bookRequest->id }}/message">(Message)</a>                                        
+                                    
+                                    
+                                    @if (auth()->check())
+                                        @if ( auth()->user()->id  != $bookRequest->user->id)
+                                            <a href="/book-requests/{{ $bookRequest->id }}/message">(Message)</a>                                        
+                                        @endif
+                                        @if ( auth()->user()->id  == $bookRequest->user->id)
+                                            <a href="book-requests/{{ $bookRequest->id }}/delete">(Delete)</a>                                        
+                                        @endif
                                     @endif
-
-                                    @if ( auth()->user()->id  == $bookRequest->user->id)
-                                        <a href="book-requests/{{ $bookRequest->id }}/delete">(Delete)</a>                                        
-                                     @endif
                                     <span class="pull-right"><small><strong>Requested by:</strong> {{ $bookRequest->user->name }}</small></span>
                                     <hr>
                                 </li>
